--- a/src/device/esp/esp8266.S
+++ b/src/device/esp/esp8266.S
@@ -1,6 +1,0 @@
-
-.section .text.tinygo_scanCurrentStack
-.global tinygo_scanCurrentStack
-tinygo_scanCurrentStack:
-    // TODO: save callee saved registers on the stack
-    j tinygo_scanstack