--- a/src/machine/board_atsame54-xpro.go
+++ b/src/machine/board_atsame54-xpro.go
@@ -4,7 +4,7 @@
 package machine
 
 import (
-	"device/sam"
+	"tinygo.org/x/device/sam"
 )
 
 // Definition for compatibility, but not used