--- a/src/machine/board_badger2040.go
+++ b/src/machine/board_badger2040.go
@@ -9,8 +9,9 @@
 package machine
 
 import (
-	"device/rp"
 	"runtime/interrupt"
+
+	"tinygo.org/x/device/rp"
 )
 
 const (