--- a/src/machine/board_feather-m4-can.go
+++ b/src/machine/board_feather-m4-can.go
@@ -4,7 +4,7 @@
 package machine
 
 import (
-	"device/sam"
+	"tinygo.org/x/device/sam"
 )
 
 // used to reset into bootloader