--- a/src/machine/board_hifive1b_baremetal.go
+++ b/src/machine/board_hifive1b_baremetal.go
@@ -3,7 +3,7 @@
 
 package machine
 
-import "device/sifive"
+import "tinygo.org/x/device/sifive"
 
 // SPI on the HiFive1.
 var (