--- a/src/machine/board_lgt92.go
+++ b/src/machine/board_lgt92.go
@@ -4,8 +4,9 @@
 package machine
 
 import (
-	"device/stm32"
 	"runtime/interrupt"
+
+	"tinygo.org/x/device/stm32"
 )
 
 const (