--- a/src/machine/board_macropad-rp2040.go
+++ b/src/machine/board_macropad-rp2040.go
@@ -4,8 +4,9 @@
 package machine
 
 import (
-	"device/rp"
 	"runtime/interrupt"
+
+	"tinygo.org/x/device/rp"
 )
 
 const (