--- a/src/machine/board_nucleol552ze.go
+++ b/src/machine/board_nucleol552ze.go
@@ -4,8 +4,9 @@
 package machine
 
 import (
-	"device/stm32"
 	"runtime/interrupt"
+
+	"tinygo.org/x/device/stm32"
 )
 
 const (