--- a/src/machine/board_pico.go
+++ b/src/machine/board_pico.go
@@ -4,8 +4,9 @@
 package machine
 
 import (
-	"device/rp"
 	"runtime/interrupt"
+
+	"tinygo.org/x/device/rp"
 )
 
 // GPIO pins