--- a/src/machine/board_teensy41.go
+++ b/src/machine/board_teensy41.go
@@ -4,8 +4,9 @@
 package machine
 
 import (
-	"device/nxp"
 	"runtime/interrupt"
+
+	"tinygo.org/x/device/nxp"
 )
 
 // Digital pins