--- a/src/machine/machine_atmega.go
+++ b/src/machine/machine_atmega.go
@@ -4,10 +4,11 @@
 package machine
 
 import (
-	"device/avr"
 	"runtime/interrupt"
 	"runtime/volatile"
 	"unsafe"
+
+	"tinygo.org/x/device/avr"
 )
 
 // I2C on AVR.