--- a/src/machine/machine_atmega1284p.go
+++ b/src/machine/machine_atmega1284p.go
@@ -4,8 +4,9 @@
 package machine
 
 import (
-	"device/avr"
 	"runtime/volatile"
+
+	"tinygo.org/x/device/avr"
 )
 
 const irq_USART0_RX = avr.IRQ_USART0_RX