--- a/src/machine/machine_atmega2560.go
+++ b/src/machine/machine_atmega2560.go
@@ -4,8 +4,9 @@
 package machine
 
 import (
-	"device/avr"
 	"runtime/volatile"
+
+	"tinygo.org/x/device/avr"
 )
 
 const irq_USART0_RX = avr.IRQ_USART0_RX