--- a/src/machine/machine_atsamd21.go
+++ b/src/machine/machine_atsamd21.go
@@ -8,11 +8,12 @@
 package machine
 
 import (
-	"device/arm"
-	"device/sam"
 	"errors"
 	"runtime/interrupt"
 	"unsafe"
+
+	"tinygo.org/x/device/arm"
+	"tinygo.org/x/device/sam"
 )
 
 const deviceName = sam.Device