--- a/src/machine/machine_atsamd21_usb.go
+++ b/src/machine/machine_atsamd21_usb.go
@@ -4,10 +4,11 @@
 package machine
 
 import (
-	"device/sam"
 	"machine/usb"
 	"runtime/interrupt"
 	"unsafe"
+
+	"tinygo.org/x/device/sam"
 )
 
 const (