--- a/src/machine/machine_atsamd21e18.go
+++ b/src/machine/machine_atsamd21e18.go
@@ -8,8 +8,9 @@
 package machine
 
 import (
-	"device/sam"
 	"runtime/interrupt"
+
+	"tinygo.org/x/device/sam"
 )
 
 var (