--- a/src/machine/machine_atsame5x_can.go
+++ b/src/machine/machine_atsame5x_can.go
@@ -4,10 +4,11 @@
 package machine
 
 import (
-	"device/sam"
 	"errors"
 	"runtime/interrupt"
 	"unsafe"
+
+	"tinygo.org/x/device/sam"
 )
 
 const (