--- a/src/machine/machine_attiny85.go
+++ b/src/machine/machine_attiny85.go
@@ -4,8 +4,9 @@
 package machine
 
 import (
-	"device/avr"
 	"runtime/volatile"
+
+	"tinygo.org/x/device/avr"
 )
 
 const (