--- a/src/machine/machine_esp8266.go
+++ b/src/machine/machine_esp8266.go
@@ -4,8 +4,9 @@
 package machine
 
 import (
-	"device/esp"
 	"runtime/volatile"
+
+	"tinygo.org/x/device/esp"
 )
 
 const deviceName = esp.Device