--- a/src/machine/machine_fe310.go
+++ b/src/machine/machine_fe310.go
@@ -4,9 +4,10 @@
 package machine
 
 import (
-	"device/sifive"
 	"runtime/interrupt"
 	"unsafe"
+
+	"tinygo.org/x/device/sifive"
 )
 
 const deviceName = sifive.Device