--- a/src/machine/machine_mimxrt1062_uart.go
+++ b/src/machine/machine_mimxrt1062_uart.go
@@ -4,9 +4,10 @@
 package machine
 
 import (
-	"device/nxp"
 	"runtime/interrupt"
 	"runtime/volatile"
+
+	"tinygo.org/x/device/nxp"
 )
 
 // UART peripheral abstraction layer for the MIMXRT1062