--- a/src/machine/machine_nrf.go
+++ b/src/machine/machine_nrf.go
@@ -4,10 +4,11 @@
 package machine
 
 import (
-	"device/nrf"
 	"errors"
 	"runtime/interrupt"
 	"unsafe"
+
+	"tinygo.org/x/device/nrf"
 )
 
 var (