--- a/src/machine/machine_nrf52840_enter_bootloader.go
+++ b/src/machine/machine_nrf52840_enter_bootloader.go
@@ -4,8 +4,8 @@
 package machine
 
 import (
-	"device/arm"
-	"device/nrf"
+	"tinygo.org/x/device/arm"
+	"tinygo.org/x/device/nrf"
 )
 
 const (