--- a/src/machine/machine_nrf528xx.go
+++ b/src/machine/machine_nrf528xx.go
@@ -4,9 +4,10 @@
 package machine
 
 import (
-	"device/nrf"
 	"runtime/volatile"
 	"unsafe"
+
+	"tinygo.org/x/device/nrf"
 )
 
 func CPUFrequency() uint32 {