--- a/src/machine/machine_nxpmk66f18.go
+++ b/src/machine/machine_nxpmk66f18.go
@@ -33,9 +33,10 @@
 package machine
 
 import (
-	"device/nxp"
 	"runtime/volatile"
 	"unsafe"
+
+	"tinygo.org/x/device/nxp"
 )
 
 const deviceName = nxp.Device