--- a/src/machine/machine_rp2040.go
+++ b/src/machine/machine_rp2040.go
@@ -4,7 +4,7 @@
 package machine
 
 import (
-	"device/rp"
+	"tinygo.org/x/device/rp"
 )
 
 const deviceName = rp.Device