--- a/src/machine/machine_rp2040_adc.go
+++ b/src/machine/machine_rp2040_adc.go
@@ -4,9 +4,10 @@
 package machine
 
 import (
-	"device/rp"
 	"errors"
 	"sync"
+
+	"tinygo.org/x/device/rp"
 )
 
 // ADCChannel is the ADC peripheral mux channel. 0-4.