--- a/src/machine/machine_rp2040_i2c.go
+++ b/src/machine/machine_rp2040_i2c.go
@@ -4,9 +4,10 @@
 package machine
 
 import (
-	"device/rp"
 	"errors"
 	"internal/itoa"
+
+	"tinygo.org/x/device/rp"
 )
 
 // I2C on the RP2040.