--- a/src/machine/machine_rp2040_pll.go
+++ b/src/machine/machine_rp2040_pll.go
@@ -4,9 +4,10 @@
 package machine
 
 import (
-	"device/rp"
 	"runtime/volatile"
 	"unsafe"
+
+	"tinygo.org/x/device/rp"
 )
 
 type pll struct {