--- a/src/machine/machine_rp2040_pwm.go
+++ b/src/machine/machine_rp2040_pwm.go
@@ -4,11 +4,12 @@
 package machine
 
 import (
-	"device/rp"
 	"errors"
 	"math"
 	"runtime/volatile"
 	"unsafe"
+
+	"tinygo.org/x/device/rp"
 )
 
 var (