--- a/src/machine/machine_rp2040_spi.go
+++ b/src/machine/machine_rp2040_spi.go
@@ -4,8 +4,9 @@
 package machine
 
 import (
-	"device/rp"
 	"errors"
+
+	"tinygo.org/x/device/rp"
 )
 
 // SPI on the RP2040