--- a/src/machine/machine_stm32.go
+++ b/src/machine/machine_stm32.go
@@ -3,7 +3,7 @@
 
 package machine
 
-import "device/stm32"
+import "tinygo.org/x/device/stm32"
 
 const deviceName = stm32.Device