--- a/src/machine/machine_stm32_adc_f1.go
+++ b/src/machine/machine_stm32_adc_f1.go
@@ -4,8 +4,9 @@
 package machine
 
 import (
-	"device/stm32"
 	"unsafe"
+
+	"tinygo.org/x/device/stm32"
 )
 
 const (