--- a/src/machine/machine_stm32_exti_syscfg_noenable.go
+++ b/src/machine/machine_stm32_exti_syscfg_noenable.go
@@ -4,8 +4,9 @@
 package machine
 
 import (
-	"device/stm32"
 	"runtime/volatile"
+
+	"tinygo.org/x/device/stm32"
 )
 
 func getEXTIConfigRegister(pin uint8) *volatile.Register32 {