--- a/src/machine/machine_stm32_gpio_revb_mp.go
+++ b/src/machine/machine_stm32_gpio_revb_mp.go
@@ -4,7 +4,7 @@
 package machine
 
 import (
-	"device/stm32"
+	"tinygo.org/x/device/stm32"
 )
 
 //