--- a/src/machine/machine_stm32_i2c_reva.go
+++ b/src/machine/machine_stm32_i2c_reva.go
@@ -7,8 +7,9 @@ package machine
 // of MCUs.
 
 import (
-	"device/stm32"
 	"unsafe"
+
+	"tinygo.org/x/device/stm32"
 )
 
 const (