--- a/src/machine/machine_stm32_i2c_revb.go
+++ b/src/machine/machine_stm32_i2c_revb.go
@@ -4,8 +4,9 @@
 package machine
 
 import (
-	"device/stm32"
 	"unsafe"
+
+	"tinygo.org/x/device/stm32"
 )
 
 //go:linkname ticks runtime.ticks