--- a/src/machine/machine_stm32_rng.go
+++ b/src/machine/machine_stm32_rng.go
@@ -3,7 +3,7 @@
 
 package machine
 
-import "device/stm32"
+import "tinygo.org/x/device/stm32"
 
 var rngInitDone = false