--- a/src/machine/machine_stm32l4.go
+++ b/src/machine/machine_stm32l4.go
@@ -4,10 +4,11 @@
 package machine
 
 import (
-	"device/stm32"
 	"runtime/interrupt"
 	"runtime/volatile"
 	"unsafe"
+
+	"tinygo.org/x/device/stm32"
 )
 
 // Peripheral abstraction layer for the stm32l4