--- a/src/machine/machine_stm32l5x2.go
+++ b/src/machine/machine_stm32l5x2.go
@@ -6,7 +6,7 @@ package machine
 // Peripheral abstraction layer for the stm32f407
 
 import (
-	"device/stm32"
+	"tinygo.org/x/device/stm32"
 )
 
 func CPUFrequency() uint32 {