--- a/src/runtime/arch_cortexm.go
+++ b/src/runtime/arch_cortexm.go
@@ -4,7 +4,7 @@
 package runtime
 
 import (
-	"device/arm"
+	"tinygo.org/x/device/arm"
 )
 
 const GOARCH = "arm"