--- a/src/runtime/runtime_atmega.go
+++ b/src/runtime/runtime_atmega.go
@@ -4,8 +4,9 @@
 package runtime
 
 import (
-	"device/avr"
 	"machine"
+
+	"tinygo.org/x/device/avr"
 )
 
 func initUART() {