--- a/src/runtime/runtime_atsamd21.go
+++ b/src/runtime/runtime_atsamd21.go
@@ -4,13 +4,14 @@
 package runtime
 
 import (
-	"device/arm"
-	"device/sam"
 	"machine"
 	"machine/usb/cdc"
 	"runtime/interrupt"
 	"runtime/volatile"
 	"unsafe"
+
+	"tinygo.org/x/device/arm"
+	"tinygo.org/x/device/sam"
 )
 
 type timeUnit int64