--- a/src/runtime/runtime_atsamd21e18.go
+++ b/src/runtime/runtime_atsamd21e18.go
@@ -4,7 +4,7 @@
 package runtime
 
 import (
-	"device/sam"
+	"tinygo.org/x/device/sam"
 )
 
 func initSERCOMClocks() {