--- a/src/runtime/runtime_atsamd51g19.go
+++ b/src/runtime/runtime_atsamd51g19.go
@@ -4,7 +4,7 @@
 package runtime
 
 import (
-	"device/sam"
+	"tinygo.org/x/device/sam"
 )
 
 func initSERCOMClocks() {