--- a/src/runtime/runtime_atsamd51j20.go
+++ b/src/runtime/runtime_atsamd51j20.go
@@ -4,7 +4,7 @@
 package runtime
 
 import (
-	"device/sam"
+	"tinygo.org/x/device/sam"
 )
 
 func initSERCOMClocks() {