@aykevl @deadprogram 
refs #2634
Unlike #2634, I was able to build `tinygo-org/drivers/examples/ili9341/pyportal_boing`.
Perhaps we are proceeding correctly.

The entire device repository is over 500MB and cannot `go get`.
For this reason, only samd21 and samd51 can be built at this time.

```
$ go get tinygo.org/x/device@e0608542069c7beaa72e203e8533de55208ad166
go: downloading tinygo.org/x/device v0.0.0-20220804091950-e0608542069c
go: tinygo.org/x/device@e0608542069c7beaa72e203e8533de55208ad166: create zip: module source tree too large (max size is 524288000 bytes)
```

Finally, I think directories such as `tinygo.org/device/sam` or `tinygo.org/device/stm32` might be a good place.
To place this, we need to associate `tinygo.org/device/sam` with `github.com/tinygo-org/sam`.

In that case, only `device/asm.go` will have trouble finding a place to put it.
We need to think about good placement.