@sago35 2022-08-04
CI is likely to be faster, as intended.

CircleCI : test-llvm14-go119

| test | before | after |
| -- | ------ | -- |
| Pull submodules | 49 sec | 65 sec |
| make gen-device -j4 | 186 sec | 0 sec |