@sago35 2022-08-04
What should I do about this?

```
=== RUN   TestCompiler
=== PAUSE TestCompiler
=== CONT  TestCompiler
=== RUN   TestCompiler/basic.go
=== RUN   TestCompiler/pointer.go
=== RUN   TestCompiler/slice.go
=== RUN   TestCompiler/string.go
=== RUN   TestCompiler/float.go
=== RUN   TestCompiler/interface.go
=== RUN   TestCompiler/func.go
=== RUN   TestCompiler/defer.go-cortex-m-qemu
Error: ../../../../.cache/tinygo/goroot-e70de3bab50987dd6e72e1df1be7fed9a7f9a954d989c7d2a929c3536bbe1629/src/runtime/interrupt/interrupt_cortexm.go:7:2: cannot find package
FAIL	github.com/tinygo-org/tinygo/compiler	0.929s
```


```
 === CONT  TestBuild/EmulatedRISCV/alias.go
Error: ../../../.cache/tinygo/goroot-e70de3bab50987dd6e72e1df1be7fed9a7f9a954d989c7d2a929c3536bbe1629/src/runtime/interrupt/interrupt_tinygoriscv.go:6:8: cannot find package
FAIL	github.com/tinygo-org/tinygo	286.048s
```