@deadprogram 2022-08-06
Just a few general comments.

As @aykevl mentioned the custom domain package naming would be `tinygo.org/x/device/sam`. We use that to more easily handle the requirements for custom domains/caching/etc.

> So for example tinygo.org/x/device/sam is in its own repository.

Each of the packages `tinygo.org/x/device/sam` could in fact be a separate package, but still within the main repo `device`. As long as each subpackage has its own `go.mod` file, they will be treated as separate, and avoid having to pull down files for families until you actually try to compile some code that needs it.