@aykevl 2022-08-31
> As long as each subpackage has its own `go.mod` file, they will be treated as separate, and avoid having to pull down files for families until you actually try to compile some code that needs it.

I have tried this, and couldn't get it to work. If someone gets it to work, I'd be happy to hear.